requirements.txt
Django>=4.0
django-ckeditor
Pillow>=11.0.0
